+++ HISTORY.rst
@@ -9,7 +9,8 @@ Contributors to this version: Trevor James Smith (:user:`Zeitsperre`), Juliette 
 Announcements
 ^^^^^^^^^^^^^
 * `xscen` is now offered as a conda package available through Anaconda.org. Refer to the installation documentation for more information. (:issue:`149`, :pull:`171`).
-* Deprecation: Release 0.6.0 of `xscen` will be the last version to support ``xscen.extract.clisops_subset``. (:pull:`182`).
+* Deprecation: Release 0.6.0 of `xscen` will be the last version to support ``xscen.extract.clisops_subset``. Use ``xscen.spatial.subset`` instead. (:pull:`182`, :pull:`184`).
+* Deprecation: The argument `region`, used in multiple functions, has been slightly reformatted. Release 0.6.0 of `xscen` will be the last version to support the old format. (:issue:`99`, :issue:`101`, :pull:`184`).
 
 New features and enhancements
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -23,6 +24,10 @@ New features and enhancements
 * Allow searching and extracting multiple frequencies for a given variable. (:issue:`168`, :pull:`170`).
 * New masking feature in ``extract_dataset``. (:issue:`180`, :pull:`182`).
 * New function ``xs.spatial.subset`` to replace ``xs.extract.clisops_subset`` and add method "sel". (:issue:`180`, :pull:`182`).
+* Add long_name attribute to diagnostics. ( :pull:`189`).
+* Added a new YAML-centric notebook (:issue:`8`, :pull:`191`).
+* New ``utils.standardize_periods`` to standardize that argument across multiple functions. (:issue:`87`, :pull:`192`).
+* New `coverage_kwargs` argument added to ``search_data_catalogs`` to allow modifying the default values of ``subset_file_coverage``. (:issue:`87`, :pull:`192`).
 
 Breaking changes
 ^^^^^^^^^^^^^^^^
@@ -31,6 +36,8 @@ Breaking changes
 * The 'datasets' dimension of the output of ``diagnostics.measures_heatmap`` is renamed 'realization'. (:pull:`167`).
 * `_subset_file_coverage` was renamed `subset_file_coverage` and moved to ``catalog.py`` to prevent circular imports. (:pull:`170`).
 * `extract_dataset` doesn't fail when a variable is in the dataset, but not `variables_and_freqs`. (:pull:`185`).
+* The argument `period`, used in multiple function, is now always a single list, while `periods` is more flexible. (:issue:`87`, :pull:`192`).
+* The parameters `reference_period` and `simulation_period` of ``xscen.train`` and ``xscen.adjust`` were renamed `period/periods` to respect the point above. (:issue:`87`, :pull:`192`).
 
 Bug fixes
 ^^^^^^^^^
@@ -41,7 +48,7 @@ Bug fixes
 * `xscen.utils.unstack_fill_nan`` can now handle datasets that have non dimension coordinates. (:issue:`156`, :pull:`175`).
 * `extract_dataset` now skips a simulation way earlier if the frequency doesn't match. (:pull:`170`).
 * `extract_dataset` now correctly tries to extract in reverse timedelta order. (:pull:`170`).
-
+* `compute_deltas` no longer creates all NaN values if the input dataset is in a non-standard calendar. (:pull:`188`).
 
 Internal changes
 ^^^^^^^^^^^^^^^^
@@ -53,6 +60,8 @@ Internal changes
 * Better warning messages in ``_subset_file_coverage`` when coverage is insufficient. (:pull:`125`).
 * The top-level Makefile now includes a `linkcheck` recipe, and the ReadTheDocs configuration no longer reinstalls the `llvmlite` compiler library. (:pull:`173`).
 * The checkups on coverage and duplicates can now be skipped in `subset_file_coverage`. (:pull:`170`).
+* Changed the `ProjectCatalog` docstrings to make it more obvious that it needs to be created empty. (:issue:`99`, :pull:`184`).
+* Added parse_config to `creep_fill`, `creep_weights`, and `reduce_ensemble` (:pull:`191`).
 
 v0.5.0 (2023-02-28)
 -------------------